multiple guest accounts
handle the situation where the user creates a guest account and then connects an eth address. then they come back and make a new guest account and again they try to connect an eth address. since we know they signed that message with their address, we disconnect the old guest account and connect their address to the new guest account.